% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print,simple.progress.bar-method}
\alias{print,simple.progress.bar-method}
\title{Print method for simple progress bars}
\usage{
\S4method{print}{simple.progress.bar}(x)
}
\arguments{
\item{x}{A simple progress bar}
}
\value{
Nothing. Prints the progress bar to the console
}
\description{
Print method for simple progress bars
}
\examples{

# Example without custom colours or custom message
pb<-create_pb(length=10, print=FALSE)
print(pb)

}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
