% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simple.progress.bar}
\alias{simple.progress.bar}
\title{Class of a simple progress bar}
\description{
Class of a simple progress bar
}
\examples{
pb<-create_pb(length=10, print=FALSE)
class(pb)
}
\author{
\enc{Telarico, Fabio Ashtar}{Fabio Ashtar Telarico}
}
